feat: Add confirmation dialog before discarding project changes, moved shipping section in summary layout
