+++ css/configurator.css
@@ -197,6 +197,12 @@ header .flex.justify-center.items-center {
   color: #111827;
 }
 
+.summary-shipping-section {
+  display: flex;
+  flex-direction: column;
+  gap: 0.75rem;
+}
+
 .summary-options-groups {
   display: flex;
   flex-direction: column;
