+++ js/main.js
@@ -733,5 +733,5 @@ if (designsSection) {
   // Log successful app load with timestamp
   console.log('%c✓ WoodLab Configurator loaded successfully', 'color: #10b981; font-weight: bold; font-size: 12px;');
   console.log('Last updated: 2026-01-05 16:31');
-  console.log('Edit ver: 385');
+  console.log('Edit ver: 386');
 });
